+++ src/simpandas/frame.py
@@ -5,8 +5,8 @@ Created on Sun Oct 11 11:14:32 2020
 @author: Martín Carlos Araya <martinaraya@gmail.com>
 """
 
-__version__ = '0.84.0'
-__release__ = 20230726
+__version__ = '0.90.0'
+__release__ = 20260420
 __all__ = ['SimDataFrame']
 
 import logging
@@ -492,46 +492,6 @@ class SimDataFrame(SimBasics, DataFrame):
     def itertuples(self, index=True, name='Pandas'):
         """Iterate over DataFrame rows as namedtuples."""
         yield from self.as_dataframe().itertuples(index=index, name=name)
-
-    def to_csv(self, path_or_buf=None, units=True, index=True, **kwargs):
-        """Write to CSV with a units row.  See ``writers.csv.write_csv``."""
-        from simpandas.writers.csv import write_csv
-        return write_csv(self, path_or_buf, units=units, index=index, **kwargs)
-
-    def to_json(self, path_or_buf=None, **kwargs):
-        """Write to JSON with units metadata.  See ``writers.json.write_json``."""
-        from simpandas.writers.json import write_json
-        return write_json(self, path_or_buf, **kwargs)
-
-    def to_hdf5(self, filepath, group='simpandas', compression='gzip'):
-        """Write to HDF5 with units metadata.  See ``writers.h5.write_hdf5``."""
-        from simpandas.writers.h5 import write_hdf5
-        return write_hdf5(self, filepath, group=group, compression=compression)
-
-    def to_summary(self, smspec_path, unsmry_path=None, startdat=None):
-        """Write to Eclipse binary summary format.  See ``writers.summary.write_summary``."""
-        from simpandas.writers.summary import write_summary
-        return write_summary(self, smspec_path, unsmry_path=unsmry_path, startdat=startdat)
-
-    def to_parquet(self, filepath, compression='snappy', **kwargs):
-        """Write to Parquet with units metadata.  See ``writers.parquet.write_parquet``."""
-        from simpandas.writers.parquet import write_parquet
-        return write_parquet(self, filepath, compression=compression, **kwargs)
-
-    def to_prodml(self, filepath, style='timeseries', facility='SimPandas'):
-        """Write to PRODML XML.  See ``writers.prodml.write_prodml``."""
-        from simpandas.writers.prodml import write_prodml
-        return write_prodml(self, filepath, style=style, facility=facility)
-
-    def to_witsml(self, filepath, well_name='', wellbore_name='', log_name='SimPandas Export'):
-        """Write to WITSML v1.4.1.1 log XML.  See ``writers.witsml.write_witsml``."""
-        from simpandas.writers.witsml import write_witsml
-        return write_witsml(self, filepath, well_name=well_name, wellbore_name=wellbore_name, log_name=log_name)
-
-    def to_resqml(self, filepath):
-        """Write to RESQML EPC package.  See ``writers.resqml.write_resqml``."""
-        from simpandas.writers.resqml import write_resqml
-        return write_resqml(self, filepath)
 
     def to_pandas(self):
         return self.to_dataframe()